// peso_lucky_admin-main/src/pages/Report/Game/data.d.ts
export interface Params {}
export interface ListItem {
  amount: number;
  bet_content_profitable_color: string;
  bet_count: number;
  bet_users: number;
  currency: number;
  game_name: string;
  game_type: number;
  prize_amount: number;
  profit_amount: number;
  profit_rate: number;
  valid_bet: number;
  total_reg_users: number;
  total_bet_users: number;
}